pyproject.toml
[project]
name = "gemini-paper-summarizer"
version = "0.7.0"
description = "Gemini Paper Summarizer"
authors = [{ name = "7shi", email = "7shi@live.jp" }]
readme = "README.md"
license = { text = "CC0 1.0 Universal (CC0 1.0) Public Domain Dedication" }
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.11.0",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.1",
]

[project.urls]
Repository = "https://github.com/7shi/gemini-paper-summarizer"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "twine>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gp_summarize"]

[project.scripts]
gp-summarize = "gp_summarize.__main__:main"
